reorganize settings into settings/ package
