+++ webapp/settings/__init__.py
@@ -1,3 +1,4 @@
 # Re-export everything from base for backward compatibility
 # Specific deployments should reference settings.base directly
+# ruff: noqa: F405
 from settings.base import *  # noqa: F403
