+++ webapp/settings/base.py
@@ -1,5 +1,8 @@
 """
-Django settings for quefairedemesobjets project.
+Django base settings for quefairedemesobjets project.
+
+Shared by all deployment profiles (dev, production, airflow, test).
+Profile-specific overrides go in dev.py / production.py / airflow.py / test_settings.py.
 
 Generated by 'django-admin startproject' using Django 4.2.2.
 
@@ -13,14 +16,9 @@ https://docs.djangoproject.com/en/4.2/ref/settings/
 from pathlib import Path
 from urllib.parse import urlparse
 
-import logging
-
 import decouple
 import dj_database_url
-import sentry_sdk
 from import_export.formats.base_formats import CSV, XLS, XLSX
-from sentry_sdk.integrations.django import DjangoIntegration
-from sentry_sdk.integrations.logging import LoggingIntegration
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
@@ -55,7 +53,6 @@ CARTE_MAX_SOLUTION_DISPLAYED = decouple.config(
 )
 DISTANCE_MAX = decouple.config("DISTANCE_MAX", cast=int, default=30000)
 
-
 NB_CORRECTION_DISPLAYED = decouple.config(
     "NB_CORRECTION_DISPLAYED", cast=int, default=100
 )
@@ -78,18 +75,12 @@ STATS = {
     ),
 }
 
-# Common settings
-# --------------
-# SECURITY WARNING: keep the secret key used in production secret!
+# Security
+# --------
 SECRET_KEY = decouple.config("SECRET_KEY")
 SECURE_REFERRER_POLICY = "strict-origin-when-cross-origin"
-# SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = decouple.config("DEBUG", default=False, cast=bool)
-STIMULUS_DEBUG = decouple.config("STIMULUS_DEBUG", default=False, cast=bool)
-POSTHOG_DEBUG = decouple.config("POSTHOG_DEBUG", default=False, cast=bool)
-BYPASS_ACTEUR_READONLY_FIELDS = decouple.config(
-    "BYPASS_ACTEUR_READONLY_FIELDS", default=False, cast=bool
-)
+# SECURE_PROXY_SSL_HEADER / USE_X_FORWARDED_HOST / USE_X_FORWARDED_PORT
+# are set in production.py (only meaningful behind a reverse proxy with HTTPS).
 BLOCK_ROBOTS = decouple.config("BLOCK_ROBOTS", default=False, cast=bool)
 ENVIRONMENT = decouple.config("ENVIRONMENT", default="development", cast=str)
 VERSION = decouple.config("CONTAINER_VERSION", default="version", cast=str)
@@ -104,13 +95,6 @@ ALLOWED_HOSTS = [
     BASE_HOST,
     *decouple.config("ALLOWED_HOSTS", default="", cast=str).split(","),
 ]
-
-# Configure Django to trust proxy headers based on BASE_URL scheme
-if urlparse(BASE_URL).scheme == "https":
-    SECURE_PROXY_SSL_HEADER = ("HTTP_X_FORWARDED_PROTO", "https")
-    USE_X_FORWARDED_HOST = True
-    USE_X_FORWARDED_PORT = True
-
 
 # Application definition
 # ----------------------
@@ -144,13 +128,11 @@ INSTALLED_APPS = [
     "djangoql",
 ]
 
-
 FORM_RENDERER = "django.forms.renderers.TemplatesSetting"
 
 # RGAA 11.10 : marque les champs facultatifs avec « (Optional) » plutôt que
 # les champs obligatoires avec un astérisque, conformément à la
-# recommandation DSFR :
-# https://www.systeme-de-design.gouv.fr/version-courante/fr/modeles/blocs-fonctionnels/formulaires#champ-obligatoire
+# recommandation DSFR
 DSFR_MARK_OPTIONAL_FIELDS = True
 
 MIDDLEWARE = [
@@ -177,78 +159,14 @@ CACHES = {
 
 X_FRAME_OPTIONS = "ALLOWALL"
 
-WITH_DJANGO_DEBUG_TOOLBAR = decouple.config(
-    "WITH_DJANGO_DEBUG_TOOLBAR", default=False, cast=bool
-)
-WITH_DJANGO_SILK = decouple.config("WITH_DJANGO_SILK", default=False, cast=bool)
-
-if DEBUG:
-    # FIXME: A CSRF error can occur locally when using HTTPS.
-    # It is not clear yet what causes it, it might be http host headers
-    # being wrongly set when using nginx. The current workaround impacts
-    # only local development and can be considered harmless, but this
-    # might be nice to remove this setting someday.
-    CSRF_TRUSTED_ORIGINS = [BASE_URL]
-    INSTALLED_APPS.extend(["django_browser_reload"])
-    MEDIA_ROOT = "media"
-    MEDIA_URL = "/media/"
-    MIDDLEWARE.extend(
-        [
-            "django_browser_reload.middleware.BrowserReloadMiddleware",
-        ]
-    )
-
-    if WITH_DJANGO_DEBUG_TOOLBAR:
-        from debug_toolbar.settings import CONFIG_DEFAULTS
-
-        INSTALLED_APPS.append("debug_toolbar")
-        MIDDLEWARE.append("debug_toolbar.middleware.DebugToolbarMiddleware")
-        DEBUG_TOOLBAR_CONFIG = {
-            "DISABLE_PANELS": (
-                "debug_toolbar.panels.request.RequestPanel",
-                "debug_toolbar.panels.staticfiles.StaticFilesPanel",
-                "debug_toolbar.panels.alerts.AlertsPanel",
-                "debug_toolbar.panels.cache.CachePanel",
-                "debug_toolbar.panels.signals.SignalsPanel",
-                "debug_toolbar.panels.community.CommunityPanel",
-                "debug_toolbar.panels.redirects.RedirectsPanel",
-                "debug_toolbar.panels.profiling.ProfilingPanel",
-            ),
-            "TOOLBAR_STORE_CLASS": "debug_toolbar.store.DatabaseStore",
-        }
-
-        def show_toolbar_callback(request):
-            path_is_not_excluded = not any(
-                p in request.path for p in patterns_to_exclude
-            )
-
-            if request.headers.get("User-Agent") == "playwright":
-                return False
-            if request.headers.get("Sec-Fetch-Dest") == "iframe":
-                # Hide in iframe
-                return False
-
-            return path_is_not_excluded
-
-        patterns_to_exclude = ["/lookbook"]
-        # ROOT_TAG_EXTRA_ATTRS is required so Turbo doesn't replace the toolbar
-        # between navigations, which would cause "panel data no longer available".
-        DEBUG_TOOLBAR_CONFIG = {
-            "SHOW_TOOLBAR_CALLBACK": show_toolbar_callback,
-            "HIDE_IN_STACKTRACES": CONFIG_DEFAULTS["HIDE_IN_STACKTRACES"]
-            + ("sentry_sdk",),
-            "ROOT_TAG_EXTRA_ATTRS": "data-turbo-permanent",
-        }
-
-    if WITH_DJANGO_SILK:
-        INSTALLED_APPS.append("silk")
-        MIDDLEWARE.append("silk.middleware.SilkyMiddleware")
-
-    CORS_ALLOW_ALL_ORIGINS = DEBUG
-
+# Logging
+# -------
+# Override via LOGLEVEL env var (default: INFO).
 
 LOGLEVEL = decouple.config(
-    "LOGLEVEL", default="info" if DEBUG else "error", cast=str
+    "LOGLEVEL",
+    default="info",
+    cast=str,
 ).upper()
 
 LOGGING = {
@@ -261,7 +179,6 @@ LOGGING = {
             "style": "{",
         },
         "default": {
-            # exact format is not important, this is the minimum information
             "format": "[%(asctime)s] [%(name)s] [%(levelname)s] : %(message)s",
         },
         "simple": {
@@ -273,7 +190,6 @@ LOGGING = {
         "console": {
             "class": "logging.StreamHandler",
             "formatter": "default",
-            "filters": ["hide_staticfiles"] if DEBUG else [],
         }
     },
     "loggers": {
@@ -290,6 +206,7 @@ LOGGING = {
         "qfdmd": {"handlers": ["console"], "level": LOGLEVEL},
     },
 }
+# In dev.py the hide_staticfiles filter is added back; in production it stays off.
 
 ROOT_URLCONF = "core.urls"
 
@@ -329,9 +246,8 @@ WSGI_APPLICATION = "core.wsgi.application"
 
 # Default primary key field type
 # https://docs.djangoproject.com/en/4.2/ref/settings/#default-auto-field
-
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
-EMAIL_BACKEND = "django.core.mail.backends.console.EmailBackend"
+# EMAIL_BACKEND — overridden in dev.py (console) and production.py (SMTP via env)
 
 
 # Database
@@ -404,14 +320,11 @@ AUTH_PASSWORD_VALIDATORS = [
     },
 ]
 
-
 LOGIN_URL = "admin:login"
 
 
 # Internationalization
 # --------------------
-# https://docs.djangoproject.com/en/4.2/topics/i18n/
-
 TIME_ZONE = "UTC"
 USE_I18N = True
 USE_TZ = True
@@ -431,7 +344,6 @@ LOCALE_PATHS = [BASE_DIR / "locale"]
 # Static files (CSS, JavaScript, Images)
 # --------------------------------------
 # https://docs.djangoproject.com/en/4.2/howto/static-files/
-
 STATIC_ROOT = BASE_DIR / "staticfiles"
 STATIC_URL = "static/"
 STATICFILES_DIRS = [
@@ -439,58 +351,17 @@ STATICFILES_DIRS = [
     BASE_DIR / "static" / "compiled",
 ]
 
-
 WHITENOISE_KEEP_ONLY_HASHED_FILES = True
-
-
-# Sentry
-# ------
-SENTRY_DSN = decouple.config("SENTRY_DSN", cast=str, default="")
-sentry_sdk.init(
-    dsn=SENTRY_DSN,
-    integrations=[
-        DjangoIntegration(),
-        LoggingIntegration(
-            level=logging.INFO,  # INFO+ log records attached as breadcrumbs
-            event_level=logging.ERROR,  # ERROR+ log records create Sentry events
-        ),
-    ],
-    environment=ENVIRONMENT,
-    # If you wish to associate users to errors (assuming you are using
-    # django.contrib.auth) you may enable sending PII data.
-    send_default_pii=False,
-    # trace only 1% of requests to get performance statistics
-    traces_sample_rate=0.01,
-    # By default the SDK will try to use the SENTRY_RELEASE
-    # environment variable, or infer a git commit
-    # SHA as release, however you may want to set
-    # something more human-readable.
-    # release="qfdmo@1.0.0",
-)
-
-
-SHELL_PLUS_PRINT_SQL = True
-
-# S3 bucket settings
-# ------------------
-AWS_ACCESS_KEY_ID = decouple.config("AWS_ACCESS_KEY_ID", default="")
-AWS_SECRET_ACCESS_KEY = decouple.config("AWS_SECRET_ACCESS_KEY", default="")
-AWS_STORAGE_BUCKET_NAME = decouple.config("AWS_STORAGE_BUCKET_NAME", default="")
-AWS_S3_REGION_NAME = decouple.config("AWS_S3_REGION_NAME", default="")
-AWS_S3_ENDPOINT_URL = decouple.config("AWS_S3_ENDPOINT_URL", default="")
 
 STORAGES = {
     "default": {
-        "BACKEND": (
-            "storages.backends.s3.S3Storage"
-            if AWS_ACCESS_KEY_ID
-            else "django.core.files.storage.FileSystemStorage"
-        ),
+        "BACKEND": "django.core.files.storage.FileSystemStorage",
     },
     "staticfiles": {
         "BACKEND": "whitenoise.storage.CompressedManifestStaticFilesStorage",
     },
 }
+# Production uses S3Storage for default — overridden in production.py.
 
 # Import / export settings
 # ------------------------
@@ -509,7 +380,6 @@ ADDRESS_SUGGESTION_FORM = decouple.config(
 UPDATE_SUGGESTION_FORM = decouple.config(
     "UPDATE_SUGGESTION_FORM", default="https://tally.so/r/3xMqd9", cast=str
 )
-
 FEEDBACK_FORM = decouple.config(
     "FEEDBACK_FORM", default="https://tally.so/r/3EW88q", cast=str
 )
@@ -540,9 +410,7 @@ SILENCED_SYSTEM_CHECKS = ["wagtailadmin.W002"]
 HOST_PROTO = decouple.config("HOST_PROTO", default="https")
 WAGTAIL_SITE_NAME = "Que faire de mes objets et déchets"
 WAGTAILADMIN_BASE_URL = BASE_URL
-# Disable autosave
 WAGTAIL_AUTOSAVE_INTERVAL = 0
-# increase ping delay : 30s
 WAGTAIL_EDITING_SESSION_PING_INTERVAL = 30000
 MODELSEARCH_BACKENDS = {
     "default": {
@@ -591,12 +459,10 @@ INSTALLED_APPS.extend(
 )
 WAGTAILIMAGES_EXTENSIONS = ["gif", "jpg", "jpeg", "png", "webp", "svg"]
 
-
 STATICFILES_FINDERS = [
     "django.contrib.staticfiles.finders.FileSystemFinder",
     "django.contrib.staticfiles.finders.AppDirectoriesFinder",
 ]
-
 
 MIDDLEWARE.extend(
     [
